Add prettier format

+++ contracts/5/Migrations.sol
@@ -3,17 +3,14 @@ pragma solidity >=0.4.22 <0.8.0;
 
 contract Migrations {
   address public owner = msg.sender;
-  uint public last_completed_migration;
+  uint256 public last_completed_migration;
 
   modifier restricted() {
-    require(
-      msg.sender == owner,
-      "This function is restricted to the contract's owner"
-    );
+    require(msg.sender == owner, "This function is restricted to the contract's owner");
     _;
   }
 
-  function setCompleted(uint completed) public restricted {
+  function setCompleted(uint256 completed) public restricted {
     last_completed_migration = completed;
   }
 }
